+++ interfaces/console/src/components/SiteConfigurationStepperDialog/index.tsx
@@ -1,11 +1,3 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * Copyright (c) 2023-present, Ukama Inc.
- */
-
 import React, { useState } from 'react';
 import {
   Dialog,
@@ -13,26 +5,18 @@ import {
   DialogContent,
   DialogActions,
   Button,
-  Box,
-  Stepper,
-  Step,
-  StepLabel,
-  TextField,
-  MenuItem,
-  Select,
-  FormControl,
-  InputLabel,
   Typography,
   Stack,
   IconButton,
+  Stepper,
+  Step,
+  StepLabel,
 } from '@mui/material';
-import { Formik, Form, Field, ErrorMessage, useField } from 'formik';
-import * as Yup from 'yup';
-import { globalUseStyles } from '@/styles/global';
-import colors from '@/theme/colors';
-import { SITE_CONFIG_STEPS } from '@/constants';
+import { Formik, Form } from 'formik';
 import CloseIcon from '@mui/icons-material/Close';
-import dynamic from 'next/dynamic';
+import { SITE_CONFIG_STEPS } from '@/constants';
+import { STEPPER_FORM_SCHEMA } from '@/helpers/formValidators';
+import SiteStepForm from '@/components/SiteStepForm';
 
 interface SiteConfigurationStepperDialogProps {
   open: boolean;
@@ -40,74 +24,9 @@ interface SiteConfigurationStepperDialogProps {
   handleFormDataSubmit: (formData: any) => void;
 }
 
-interface CustomTextFieldProps {
-  label: string;
-  name: string;
-  onChange?: (event: any) => void;
-}
-
-const CustomTextField: React.FC<CustomTextFieldProps> = ({
-  label,
-  name,
-  onChange,
-}) => {
-  const [field, meta] = useField(name);
-  const gclasses = globalUseStyles();
-
-  return (
-    <TextField
-      {...field}
-      label={label}
-      fullWidth
-      onChange={(e) => {
-        field.onChange(e);
-        if (onChange) {
-          onChange(e);
-        }
-      }}
-      InputLabelProps={{
-        shrink: true,
-      }}
-      helperText={meta.touched && meta.error ? meta.error : ''}
-      error={meta.touched && Boolean(meta.error)}
-      spellCheck={false}
-      InputProps={{
-        classes: {
-          input: gclasses.inputFieldStyle,
-        },
-      }}
-    />
-  );
-};
-
-const validationSchema = [
-  Yup.object().shape({
-    switch: Yup.string().required('Switch is required'),
-    power: Yup.string().required('Power is required'),
-    backhaul: Yup.string().required('Backhaul is required'),
-    spectrumBand: Yup.string().required('Spectrum Band is required'),
-  }),
-  Yup.object().shape({
-    siteName: Yup.string().required('Site Name is required'),
-    network: Yup.string().required('Network is required'),
-    latitude: Yup.number()
-      .required('Latitude is required')
-      .min(-90, 'Latitude must be between -90 and 90')
-      .max(90, 'Latitude must be between -90 and 90'),
-    longitude: Yup.number()
-      .required('Longitude is required')
-      .min(-180, 'Longitude must be between -180 and 180')
-      .max(180, 'Longitude must be between -180 and 180'),
-  }),
-];
-
 const SiteConfigurationStepperDialog: React.FC<
   SiteConfigurationStepperDialogProps
 > = ({ open, handleClose, handleFormDataSubmit }) => {
-  const SiteMapComponent = dynamic(() => import('../SiteMapComponent'), {
-    loading: () => <p>Site map is loading</p>,
-    ssr: false,
-  });
   const [activeStep, setActiveStep] = useState(0);
   const [lat, setLat] = useState<number>(0);
   const [lng, setLng] = useState<number>(0);
@@ -137,7 +56,7 @@ const SiteConfigurationStepperDialog: React.FC<
 
   const handleNext = async (values: any, helpers: any) => {
     try {
-      await validationSchema[activeStep].validate(values, {
+      await STEPPER_FORM_SCHEMA[activeStep].validate(values, {
         abortEarly: false,
       });
       if (activeStep === SITE_CONFIG_STEPS.length - 1) {
@@ -158,196 +77,67 @@ const SiteConfigurationStepperDialog: React.FC<
   const handleAddressChange = (address: string) => {
     setLocation(address);
   };
-  const stepContent = (step: number) => {
-    switch (step) {
-      case 0:
-        return (
-          <Box
-            component="form"
-            style={{ display: 'flex', flexDirection: 'column', gap: 16 }}
-          >
-            <Box sx={{ mt: 2, mb: 2 }}>
-              <Typography>
-                You have successfully installed your site, and need to configure
-                it. Please note that if your power or backhaul choice is
-                “other”, it can’t be monitored within Ukama’s Console.
-              </Typography>
-            </Box>
-
-            <FormControl fullWidth>
-              <InputLabel id="switch-label">Switch</InputLabel>
-              <Field
-                as={Select}
-                labelId="switch-label"
-                name="switch"
-                label="Switch"
-              >
-                <MenuItem value="8 port switch">8 port switch</MenuItem>
-                <MenuItem value="16 port switch">16 port switch</MenuItem>
-              </Field>
-
-              <ErrorMessage name="switch" component="div">
-                {(msg) => <div style={{ color: colors.red }}>{msg}</div>}
-              </ErrorMessage>
-            </FormControl>
-
-            <FormControl fullWidth>
-              <InputLabel id="power-label">Power</InputLabel>
-              <Field
-                as={Select}
-                labelId="power-label"
-                name="power"
-                label="Power"
-              >
-                <MenuItem value="Battery">Battery</MenuItem>
-                <MenuItem value="AC Power">AC Power</MenuItem>
-              </Field>
-
-              <ErrorMessage name="power" component="div">
-                {(msg) => <div style={{ color: colors.red }}>{msg}</div>}
-              </ErrorMessage>
-            </FormControl>
-
-            <FormControl fullWidth>
-              <InputLabel id="backhaul-label">Backhaul</InputLabel>
-              <Field
-                as={Select}
-                labelId="backhaul-label"
-                name="backhaul"
-                label="Backhaul"
-              >
-                <MenuItem value="ViaSAT">ViaSAT</MenuItem>
-                <MenuItem value="Other">Other</MenuItem>
-              </Field>
-
-              <ErrorMessage name="backhaul" component="div">
-                {(msg) => <div style={{ color: colors.red }}>{msg}</div>}
-              </ErrorMessage>
-            </FormControl>
-
-            <FormControl fullWidth>
-              <InputLabel id="spectrumBand-label">Spectrum Band</InputLabel>
-              <Field
-                as={Select}
-                labelId="spectrumBand-label"
-                name="spectrumBand"
-                label="Spectrum Band"
-              >
-                <MenuItem value="Band 40">Band 40</MenuItem>
-                <MenuItem value="Band 41">Band 41</MenuItem>
-              </Field>
-
-              <ErrorMessage name="spectrumBand" component="div">
-                {(msg) => <div style={{ color: colors.red }}>{msg}</div>}
-              </ErrorMessage>
-            </FormControl>
-          </Box>
-        );
-      case 1:
-        return (
-          <Box
-            component="form"
-            style={{ display: 'flex', flexDirection: 'column', gap: 16 }}
-          >
-            <Box sx={{ mt: 2, mb: 2 }}>
-              <Typography>
-                Please name your site for your ease of reference, and assign it
-                to a network.
-              </Typography>
-            </Box>
-            <SiteMapComponent
-              posix={[lat, lng]}
-              onAddressChange={handleAddressChange}
-            />
-
-            <Box>
-              <Stack direction="column" spacing={1} justifyItems={'center'}>
-                <Typography
-                  variant="body2"
-                  sx={{ color: `${colors.darkGray}` }}
-                >
-                  LOCATION
-                </Typography>
-                <Typography variant="body2" color="initial">
-                  {location || 'Fetching site location...'}
-                </Typography>
-              </Stack>
-            </Box>
-            <CustomTextField
-              label="Longitude"
-              name="longitude"
-              onChange={handleLngChange}
-            />
-            <CustomTextField
-              label="Latitude"
-              name="latitude"
-              onChange={handleLatChange}
-            />
-            <CustomTextField label="Site Name" name="siteName" />
-            <CustomTextField label="Network" name="network" />
-          </Box>
-        );
-      default:
-        return 'Unknown step';
-    }
-  };
 
   return (
-    <Dialog open={open} onClose={handleClose} maxWidth="md" fullWidth>
+    <Dialog
+      open={open}
+      onClose={handleClose}
+      maxWidth="sm"
+      fullWidth
+      aria-labelledby="site-config-dialog-title"
+    >
       <DialogTitle>
-        Site Configuration
+        <Typography variant="h6" color="initial">
+          Site Configuration
+        </Typography>
         <IconButton
           aria-label="close"
           onClick={handleClose}
-          style={{ position: 'absolute', right: '8px', top: '8px' }}
+          sx={{ position: 'absolute', right: 8, top: 8 }}
         >
           <CloseIcon />
         </IconButton>
       </DialogTitle>
       <Formik
         initialValues={initialValues}
-        validationSchema={validationSchema[activeStep]}
-        onSubmit={handleNext}
+        validationSchema={STEPPER_FORM_SCHEMA[activeStep]}
+        onSubmit={(values, helpers) => handleNext(values, helpers)}
       >
-        {({ handleSubmit, isSubmitting }) => (
+        {({ handleSubmit }) => (
           <Form onSubmit={handleSubmit}>
             <DialogContent>
-              <Stepper activeStep={activeStep}>
+              <Stepper activeStep={activeStep} alternativeLabel>
                 {SITE_CONFIG_STEPS.map((label) => (
                   <Step key={label}>
                     <StepLabel>{label}</StepLabel>
                   </Step>
                 ))}
               </Stepper>
-              {stepContent(activeStep)}
+              <SiteStepForm
+                step={activeStep}
+                lat={lat}
+                lng={lng}
+                location={location}
+                handleLatChange={handleLatChange}
+                handleLngChange={handleLngChange}
+                handleAddressChange={handleAddressChange}
+              />
             </DialogContent>
             <DialogActions>
-              <Button
-                onClick={() => {
-                  handleClose(), setActiveStep(0);
-                }}
-                color="secondary"
-              >
-                Cancel
-              </Button>
-              {activeStep === SITE_CONFIG_STEPS.length - 1 ? (
-                <Button type="submit" color="primary" disabled={isSubmitting}>
-                  Save
+              {activeStep > 0 && (
+                <Button
+                  onClick={handleBack}
+                  variant="contained"
+                  color="primary"
+                >
+                  Back
                 </Button>
-              ) : (
-                <>
-                  <Button
-                    onClick={handleBack}
-                    color="primary"
-                    disabled={activeStep === 0 || isSubmitting}
-                  >
-                    Back
-                  </Button>
-                  <Button type="submit" color="primary" disabled={isSubmitting}>
-                    Next
-                  </Button>
-                </>
               )}
+              <Button type="submit" variant="contained" color="primary">
+                {activeStep === SITE_CONFIG_STEPS.length - 1
+                  ? 'Finish'
+                  : 'Next'}
+              </Button>
             </DialogActions>
           </Form>
         )}
