+++ interfaces/console/src/helpers/formValidators.ts
@@ -7,15 +7,15 @@
  */
 
 import * as Yup from 'yup';
+
+// Validation rules
 const nameValidationRule = Yup.string().required('Name is required.');
 const networkNameValidationRule = Yup.string().required(
   'Network name is required.',
 );
-
 const emailValidatonRule = Yup.string()
   .required('Email is required.')
   .email('Please enter a valid email.');
-
 const iccidValidator = Yup.string()
   .notRequired()
   .nullable()
@@ -25,20 +25,46 @@ const iccidValidator = Yup.string()
 const securitycodeValidator = Yup.string().required(
   'Security code is required.',
 );
-const ESIM_FORM_SCHEMA = {
+
+// Stepper form validation schema
+const STEPPER_FORM_SCHEMA = [
+  Yup.object().shape({
+    switch: Yup.string().required('Switch is required'),
+    power: Yup.string().required('Power is required'),
+    backhaul: Yup.string().required('Backhaul is required'),
+    spectrumBand: Yup.string().required('Spectrum Band is required'),
+  }),
+  Yup.object().shape({
+    siteName: Yup.string().required('Site Name is required'),
+    network: Yup.string().required('Network is required'),
+    latitude: Yup.number()
+      .required('Latitude is required')
+      .min(-90, 'Latitude must be between -90 and 90')
+      .max(90, 'Latitude must be between -90 and 90'),
+    longitude: Yup.number()
+      .required('Longitude is required')
+      .min(-180, 'Longitude must be between -180 and 180')
+      .max(180, 'Longitude must be between -180 and 180'),
+  }),
+];
+
+// Validation schemas
+const ESIM_FORM_SCHEMA = Yup.object().shape({
   email: emailValidatonRule,
   name: nameValidationRule,
   simiccid: iccidValidator,
-};
-const NETWORK_NAME_SCHEMA_VALIDATOR = {
+});
+const NETWORK_NAME_SCHEMA_VALIDATOR = Yup.object().shape({
   name: networkNameValidationRule,
-};
-const PHYSICAL_SIM_FORM_SCHEMA = {
+});
+const PHYSICAL_SIM_FORM_SCHEMA = Yup.object().shape({
   iccid: iccidValidator,
   securityCode: securitycodeValidator,
-};
+});
+
 export {
   ESIM_FORM_SCHEMA,
   NETWORK_NAME_SCHEMA_VALIDATOR,
   PHYSICAL_SIM_FORM_SCHEMA,
+  STEPPER_FORM_SCHEMA,
 };
